move score computation to collections package
